feat: migrate RpcClientManager / RpcServerManager to not be tightly coupled to RTCEngine
