+++ src/room/rpc/index.ts
@@ -1,8 +1,10 @@
 // SPDX-FileCopyrightText: 2026 LiveKit, Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
-export { default as RpcClientManager } from './RpcClientManager';
-export { default as RpcServerManager } from './RpcServerManager';
+export { default as RpcClientManager } from './client/RpcClientManager';
+export type { RpcClientManagerCallbacks } from './client/events';
+export { default as RpcServerManager } from './server/RpcServerManager';
+export type { RpcServerManagerCallbacks } from './server/events';
 export {
   type PerformRpcParams,
   RPC_DATA_STREAM_TOPIC,
