+++ src/room/rpc/server/events.ts
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: 2026 LiveKit, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+import type { DataPacket, DataPacket_Kind } from '@livekit/protocol';
+
+export type EventSendDataPacket = {
+  packet: DataPacket;
+  kind: DataPacket_Kind;
+};
+
+export type RpcServerManagerCallbacks = {
+  sendDataPacket: (event: EventSendDataPacket) => void;
+};
